Fix "File exists" error on startup
When booting the container, an `mkdir: can't create directory '/data': File exists` error is printed to the logs. Irritatingly, nothing gets printed after this error for a while since the chown/chmod calls may take a while to complete, so the user may get the impression the container failed to start up.